# Set stackTraceLimit to 0 in fileSystemEntryExists